Week2/requirements.txt
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
idna==3.4
imageio==2.31.5
imutils==0.5.4
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-metrics==0.1.4
mpmath==1.3.0
ninja==1.11.1.1
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pyclipper==1.3.0.post5
pyparsing==3.1.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
